refactor: move DDL gen to `bigframes.session._io`, add missing test
